requirements.txt
alembic==1.6.5
blinker==1.6.2
click==8.1.6
factory-boy==3.3.0
Faker==19.2.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dateutil==2.8.1
python-dotenv==1.0.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.3.6
